Reorganized files inside directories
