--- js/client_utilities.js
@@ -1,3 +1,0 @@
-(function (e) {
-
-})(jQuery);
